add Step3

+++ Lesson20/src/main/java/YA/Lesson20/DAO/UserRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepo extends JpaRepository<User,Long> {
+public interface UserRepo extends JpaRepository<User,Integer> {
 
     public Optional<User> findByEmail(String email);
